+++ docs/conf.py
@@ -25,7 +25,7 @@ from pyrf.version import __version__
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.graphviz']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.graphviz', 'rst2pdf.pdfbuilder']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -206,6 +206,7 @@ latex_documents = [
 # If false, no module index is generated.
 #latex_domain_indices = True
 
+pdf_documents = [('index', u'rst2pdf', u'pyRF API Documentation', u'ThinkRF')]
 
 # -- Options for manual page output --------------------------------------------